Added methodFormatMarkdownTables(...)

+++ DocConversion/Program.cs
@@ -4,6 +4,9 @@ using System.Text.RegularExpressions;
 
 namespace DocConversion
 {
+    /// <summary>
+    /// Represents the main program class.
+    /// </summary>
     internal partial class Program
     {
         #region Class-Constructors
@@ -37,7 +40,24 @@ namespace DocConversion
         internal static ConsoleColor ForegroundColor { get; set; } = Console.ForegroundColor;
         #endregion Properties
 
+        static void Main(string[] args)
+        {
+            if (args.Length > 0 && Directory.Exists(args[0]))
+            {
+                SourcePath = args[0];
+                TargetPath = args[0];
+            }
+            if (args.Length > 1 && Directory.Exists(args[1]))
+            {
+                TargetPath = args[1];
+            }
+            RunApp();
+        }
+
         #region App methods
+        /// <summary>
+        /// Runs the application and allows the user to interactively choose source and target paths, as well as perform file conversions.
+        /// </summary>
         public static void RunApp()
         {
             var running = false;
@@ -128,6 +148,11 @@ namespace DocConversion
                 }
             } while (running);
         }
+        /// <summary>
+        /// Prints the header information for the document conversion.
+        /// </summary>
+        /// <param name="sourcePath">The path of the source document.</param>
+        /// <param name="targetPath">The path of the target document.</param>
         private static void PrintHeader(string sourcePath, string targetPath)
         {
             var mnuIdx = 0;
@@ -155,43 +180,55 @@ namespace DocConversion
             Console.WriteLine($"[{mnuIdx++,2}] Refresh documents");
             Console.WriteLine();
         }
+        /// <summary>
+        /// Converts a source file to a target format and saves it to the specified target path.
+        /// </summary>
+        /// <param name="sourceFile">The path of the source file to be converted.</param>
+        /// <param name="targetPath">The path where the converted file will be saved.</param>
+        /// <param name="targetExtension">The extension of the target file format.</param>
         private static void ConversionTo(string sourceFile, string targetPath, string targetExtension)
         {
             var sourceExtension = Path.GetExtension(sourceFile);
             var targetFileName = $"{Path.GetFileNameWithoutExtension(sourceFile)}{targetExtension}";
 
             targetPath = Path.Combine(targetPath!, Path.GetFileNameWithoutExtension(targetFileName));
-            if (Path.Exists(targetPath))
+
+            var targetFile = Path.Combine(targetPath, targetFileName);
+
+            if (File.Exists(targetFile) && targetExtension.Equals(".md", StringComparison.CurrentCultureIgnoreCase))
             {
-                Directory.Delete(targetPath, true);
+                CleanupMarkdown(targetFile);
             }
-            if (string.IsNullOrWhiteSpace(targetPath) == false)
+            else if (File.Exists(sourceFile) && sourceExtension.Equals(".md", StringComparison.CurrentCultureIgnoreCase))
             {
-                if (sourceExtension.Equals(".md", StringComparison.CurrentCultureIgnoreCase))
+                CleanupMarkdown(sourceFile);
+            }
+            else if (File.Exists(sourceFile))
+            {
+                if (Path.Exists(targetPath))
                 {
-                    CleanupMarkdown(sourceFile);
+                    Directory.Delete(targetPath, true);
                 }
-                else
+                Directory.CreateDirectory(targetPath);
+
+                try
                 {
-                    var targetFile = Path.Combine(targetPath, targetFileName);
+                    var document = new Document(sourceFile);
 
-                    Directory.CreateDirectory(targetPath);
-
-                    try
-                    {
-                        var document = new Document(sourceFile);
-
-                        document.Save(targetFile, SaveFormat.Markdown);
-                    }
-                    catch (Exception ex)
-                    {
-                        Console.WriteLine(ex.Message);
-                    }
-
-                    CleanupMarkdown(targetFile);
+                    document.Save(targetFile, SaveFormat.Markdown);
                 }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
+
+                CleanupMarkdown(targetFile);
             }
         }
+        /// <summary>
+        /// Cleans up a Markdown file by removing unnecessary lines and replacing image tags with a generic placeholder.
+        /// </summary>
+        /// <param name="filePath">The path to the Markdown file.</param>
         private static void CleanupMarkdown(string filePath)
         {
             var targetPath = Path.GetDirectoryName(filePath);
@@ -206,7 +243,7 @@ namespace DocConversion
 
                 foreach (var line in oldLines)
                 {
-                    if (line.StartsWith("**") && line.EndsWith("**"))
+                    if (line.StartsWith("**") && line.EndsWith("**") && line.Contains("Aspose.Words", StringComparison.CurrentCultureIgnoreCase))
                     {
                         ignoreLine = true;
                     }
@@ -220,11 +257,8 @@ namespace DocConversion
                     }
                     if (ignoreLine == false)
                     {
-                        if (prevLine == default)
-                        {
-                            prevLine = line;
-                        }
-                        else if (line.StartsWith("#"))
+                        prevLine ??= line;
+                        if (line.StartsWith("#"))
                         {
                             if (string.IsNullOrWhiteSpace(prevLine) == false)
                             {
@@ -245,8 +279,8 @@ namespace DocConversion
                     }
                 }
 
-                var text = newLines.Aggregate(new StringBuilder(), (sb, l) => sb.AppendLine(l))
-                                   .ToString();
+                var formatedTableLines = FormatMarkdownTables(newLines);
+                var text = formatedTableLines.Aggregate(new StringBuilder(), (sb, l) => sb.AppendLine(l)).ToString();
                 string pattern = @"!\[[^\]]*\]";
                 string replacement = "![Illustration]";
                 string result = Regex.Replace(text, pattern, replacement);
@@ -254,6 +288,76 @@ namespace DocConversion
                 File.WriteAllText(filePath, result);
             }
         }
+
+        /// <summary>
+        /// Formats the markdown tables in the given collection of lines.
+        /// </summary>
+        /// <param name="lines">The collection of lines containing markdown tables.</param>
+        /// <returns>A collection of formatted lines with markdown tables.</returns>
+        private static IEnumerable<string> FormatMarkdownTables(IEnumerable<string> lines)
+        {
+            var result = new List<string>();
+            var tableRows = new List<string>();
+
+            foreach (var line in lines)
+            {
+                if (line.TrimStart(' ', '\t').StartsWith('|'))
+                {
+                    tableRows.Add(line);
+                }
+                else
+                {
+                    result.AddRange(FormatMarkdownTable(tableRows));
+                    result.Add(line);
+                    tableRows.Clear();
+                }
+            }
+            if (tableRows.Count > 0)
+            {
+                result.AddRange(FormatMarkdownTable(tableRows));
+            }
+            return result;
+        }
+        /// <summary>
+        /// Formats a markdown table by aligning the columns and adjusting the column widths.
+        /// </summary>
+        /// <param name="lines">The lines of the markdown table.</param>
+        /// <returns>The formatted markdown table.</returns>
+        private static IEnumerable<string> FormatMarkdownTable(IEnumerable<string> lines)
+        {
+            var result = new List<string>();
+            var table = lines.Select(r => r.Split('|', StringSplitOptions.RemoveEmptyEntries)).ToArray();
+
+            if (table.Length > 0)
+            {
+                var columnWidths = new int[table[0].Length];
+                for (var i = 0; i < table[0].Length; i++)
+                {
+                    columnWidths[i] = table.Max(r => r[i].Length);
+                }
+                foreach (var row in table)
+                {
+                    var line = "|";
+
+                    for (var i = 0; i < row.Length; i++)
+                    {
+                        line += $"{row[i].PadRight(columnWidths[i])}|";
+                    }
+                    result.Add(line);
+                }
+            }
+            else
+            {
+                result.AddRange(lines);
+            }
+            return result;
+        }
+        /// <summary>
+        /// Retrieves a list of documents from the specified path that match the given search pattern.
+        /// </summary>
+        /// <param name="path">The directory path to search for documents.</param>
+        /// <param name="searchPattern">The search pattern to match against document filenames.</param>
+        /// <returns>An enumerable collection of document file paths.</returns>
         private static IEnumerable<string> GetDocuments(string path, string searchPattern)
         {
             var result = new List<string>();
@@ -264,12 +368,20 @@ namespace DocConversion
             result.AddRange(files);
             return result;
         }
+        /// <summary>
+        /// Retrieves a collection of documents from the specified path.
+        /// </summary>
+        /// <param name="path">The path to the directory containing the documents.</param>
+        /// <returns>A collection of document paths.</returns>
         private static IEnumerable<string> GetDocuments(string path)
         {
             return GetDocuments(path, "*.*");
         }
         #endregion App methods
 
+        /// <summary>
+        /// Prints a busy progress indicator on the console.
+        /// </summary>
         internal static void PrintBusyProgress()
         {
             if (RunBusyProgress == false)
@@ -296,18 +408,5 @@ namespace DocConversion
             }
         }
 
-        static void Main(string[] args)
-        {
-            if (args.Length > 0 && Directory.Exists(args[0]))
-            {
-                SourcePath = args[0];
-                TargetPath = args[0];
-            }
-            if (args.Length > 1 && Directory.Exists(args[1]))
-            {
-                TargetPath = args[1];
-            }
-            RunApp();
-        }
     }
 }
